WIP: refactoring ThesisList
